Solid helper methods refactoring
Simplification and generalisation of some utility methods related to solid profile and managemente. Eagger to continue this line of improvements.
